chore: simpler rollup setup
