--- configs/notificare-rollup/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@notificare/rollup",
-  "private": "true",
-  "version": "0.0.0",
-  "type": "module",
-  "main": "index.js",
-  "dependencies": {
-    "@rollup/plugin-commonjs": "^25.0.4",
-    "@rollup/plugin-json": "^6.0.0",
-    "@rollup/plugin-node-resolve": "^15.0.1",
-    "@rollup/plugin-terser": "^0.4.3",
-    "@rollup/plugin-typescript": "^11.0.0",
-    "rollup": "^3.10.0",
-    "rollup-plugin-svg-import": "^2.1.0"
-  }
-}
