+++ packages/notificare-assets/rollup.config.mjs
@@ -1,5 +1,5 @@
 import typescript from 'typescript';
-import { build } from '@notificare/rollup';
+import { build } from '../../scripts/rollup/rollup.base.config.mjs';
 import pkg from './package.json' assert { type: 'json' };
 
 export default build(pkg, { typescript });
