+++ packages/notificare-cloud-api/package.json
@@ -25,7 +25,6 @@
   },
   "devDependencies": {
     "@notificare/eslint-config": "*",
-    "@notificare/rollup": "*",
     "typescript": "^5.1.6"
   },
   "publishConfig": {
