+++ packages/notificare-core/package.json
@@ -26,7 +26,6 @@
   ],
   "devDependencies": {
     "@notificare/eslint-config": "*",
-    "@notificare/rollup": "*",
     "typescript": "^5.1.6"
   },
   "publishConfig": {
