+++ packages/notificare-geo/package.json
@@ -27,7 +27,6 @@
   },
   "devDependencies": {
     "@notificare/eslint-config": "*",
-    "@notificare/rollup": "*",
     "@types/google.maps": "^3.52.6",
     "typescript": "^5.1.6"
   },
