+++ packages/notificare-inbox/package.json
@@ -27,7 +27,6 @@
   },
   "devDependencies": {
     "@notificare/eslint-config": "*",
-    "@notificare/rollup": "*",
     "typescript": "^5.1.6"
   },
   "publishConfig": {
