+++ packages/notificare-push-ui/package.json
@@ -32,7 +32,6 @@
   },
   "devDependencies": {
     "@notificare/eslint-config": "*",
-    "@notificare/rollup": "*",
     "@types/google.maps": "^3.52.6",
     "typescript": "^5.1.6"
   },
