+++ packages/notificare-ui/package.json
@@ -28,7 +28,6 @@
   },
   "devDependencies": {
     "@notificare/eslint-config": "*",
-    "@notificare/rollup": "*",
     "typescript": "^5.1.6"
   },
   "publishConfig": {
