+++ packages/notificare/rollup.config.mjs
@@ -1,5 +1,5 @@
 import typescript from 'typescript';
-import { buildUmbrella } from '@notificare/rollup';
+import { build } from '../../scripts/rollup/rollup.umbrella.config.mjs';
 import pkg from './package.json' assert { type: 'json' };
 
-export default buildUmbrella(pkg, { typescript, rootDirectory: import.meta.url });
+export default build(pkg, { typescript, rootDirectory: import.meta.url });
