+++ scripts/rollup/rollup.umbrella.config.mjs
@@ -3,120 +3,11 @@ import json from "@rollup/plugin-json";
 import terser from "@rollup/plugin-terser";
 import nodeResolve from "@rollup/plugin-node-resolve";
 import commonjs from "@rollup/plugin-commonjs";
-import svg from "rollup-plugin-svg-import";
-import { emitModulePackageFile } from "./plugins/emit-module-package-file.js";
+import emitModulePackageFile from "./plugins/rollup-emit-module-package-file.mjs";
 import { dirname, resolve } from "node:path";
 import { readFileSync } from "node:fs";
 
 export function build(pkg, options) {
-  const { typescript } = options;
-
-  const dependencies = [
-    ...Object.keys({ ...pkg.dependencies, ...pkg.peerDependencies }),
-  ];
-
-  /**
-   * CommonJS builds
-   */
-  const commonJsBuildPlugins = [
-    tsc({
-      typescript: typescript,
-      target: "ES6",
-      // noEmitOnError: true,
-    }),
-    json(),
-    svg(),
-  ];
-
-  const commonJsBuilds = [
-    {
-      input: "src/index.ts",
-      output: [
-        {
-          file: pkg.main,
-          format: "cjs",
-          sourcemap: true,
-        },
-      ],
-      external: (dependency) => dependencies.some((d) => dependency === d),
-      plugins: commonJsBuildPlugins,
-    },
-    ...(options.extraBuilds?.cjs ?? []).map((build) => ({
-      input: build.input,
-      output: [
-        {
-          file: build.output,
-          format: "cjs",
-          sourcemap: true,
-        },
-      ],
-      external: (dependency) => dependencies.some((d) => dependency === d),
-      plugins: commonJsBuildPlugins,
-    })),
-  ];
-
-  /**
-   * ESM builds
-   */
-  const esmBuildPlugins = [
-    tsc({
-      typescript: typescript,
-      target: "es2017",
-      // noEmitOnError: true,
-    }),
-    json({ preferConst: true }),
-    svg(),
-  ];
-
-  const esmBuilds = [
-    // {
-    //   input: "src/index.ts",
-    //   output: [
-    //     {
-    //       file: pkg.esm5,
-    //       format: "es",
-    //       sourcemap: true,
-    //     },
-    //   ],
-    //   external: (dependency) => dependencies.some((d) => dependency === d),
-    //   plugins: [
-    //     typescriptPlugin({
-    //       target: "es5",
-    //     }),
-    //     json(),
-    //     emitModulePackageFile(),
-    //   ],
-    // },
-    {
-      input: "src/index.ts",
-      output: [
-        {
-          file: pkg.browser,
-          format: "es",
-          sourcemap: true,
-        },
-      ],
-      external: (dependency) => dependencies.some((d) => dependency === d),
-      plugins: [...esmBuildPlugins, emitModulePackageFile()],
-    },
-    ...(options.extraBuilds?.esm ?? []).map((build) => ({
-      input: build.input,
-      output: [
-        {
-          file: build.output,
-          format: "es",
-          sourcemap: true,
-        },
-      ],
-      external: (dependency) => dependencies.some((d) => dependency === d),
-      plugins: [...esmBuildPlugins, emitModulePackageFile()],
-    })),
-  ];
-
-  return [...commonJsBuilds, ...esmBuilds];
-}
-
-export function buildUmbrella(pkg, options) {
   const { typescript, rootDirectory } = options;
 
   const externals = Object.keys(pkg.dependencies || {});
